pyproject.toml
[project]
name = "document-translator"
version = "0.1.0"
description = "Translate Chinese Calculus textbooks to English using Gemini or Ollama"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-docx>=1.1.0",
    "python-dotenv>=1.0.0",
]
